formatLong no longer corrupts covariates when factors present; unit test added
When using `csvToUMF` to convert long data.frames (i.e., `long = TRUE`), the presence of a factor variable resulted in all other variables converting to factors.  This is now fixed. Additionally, the function nows tries to automatically identify and extract site covariates and add them to the `UnmarkedFrameOccu`. However, the `reshape::recast` function used in many of these `format` functions does not play well with factors, and it may be necessary to rethink this approach moving forward.  For example, if the first variable `recast` encounters is a factor, bad things happen without an easy solution other than reorganizing your `data.frame` so that's not the case.