allow import of dataframes with multiple classes
Tweaks unmarkedFrame constructor so that obsCovs does not have to have matrix or dataframe as its *first* listed class, as long as it is a matrix or dataframe. If you decide not to do this to avoid a bunch of work accommodating tidyverse objects (I bet this is not the only place where this kind of thing comes up), you might consider including some advice in the documentation that would guide users towards the kind of workaround that I came up with e.g. using as.data.frame() to ensure that data.frame is the only class for the object provided to obsCovs. Closes #111